requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.26.0
pillow>=10.0.0
addict>=2.4.0
matplotlib>=3.7.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Frontend
streamlit>=1.28.0

# Utilities
requests>=2.31.0
python-multipart>=0.0.6

# AWS SDK (for potential S3 integration)
boto3>=1.28.0
